Move ScanCommand to commoncode.cliutils #2233
Also rename ScanCommand to GroupedHelpCommand
