+++ tests/commoncode/test_cliutils.py
@@ -38,7 +38,7 @@ from click.testing import CliRunner
 from commoncode.testcase import FileDrivenTesting
 from commoncode.cliutils import fixed_width_file_name
 from plugincode import PluggableCommandLineOption
-from scancode.cli import ScanCommand
+from scancode.cli import GroupedHelpCommand
 
 
 class TestUtils(FileDrivenTesting):
@@ -125,13 +125,13 @@ class TestFixedWidthFilename(FileDrivenTesting):
         assert expected == test
 
 
-class TestHelpGroups(FileDrivenTesting):
+class TestGroupedHelpCommand(FileDrivenTesting):
 
     test_data_dir = os.path.join(os.path.dirname(__file__), 'data')
 
-    def test_scan_help_group_and_sort_order_without_custom_class(self):
+    def test_GroupedHelpCommand_help_group_and_sort_order_without_custom_class(self):
 
-        @click.command(name='scan', cls=ScanCommand)
+        @click.command(name='scan', cls=GroupedHelpCommand)
         @click.option('--opt', is_flag=True, help='Help text for option')
         def scan(opt):
             pass
@@ -142,9 +142,9 @@ class TestHelpGroups(FileDrivenTesting):
         assert MISC_GROUP in result.output
         assert  '--opt   Help text for option' in result.output
 
-    def test_scan_help_group_and_sort_order_with_custom_class(self):
+    def test_GroupedHelpCommand_with_help_group_and_sort_order_with_custom_class(self):
 
-        @click.command(name='scan', cls=ScanCommand)
+        @click.command(name='scan', cls=GroupedHelpCommand)
         @click.option('--opt', is_flag=True, sort_order=10,
                       help='Help text for option', cls=PluggableCommandLineOption)
         def scan(opt):
@@ -155,10 +155,10 @@ class TestHelpGroups(FileDrivenTesting):
         from plugincode import MISC_GROUP
         assert MISC_GROUP + ':\n    --opt   Help text for option\n' in result.output
 
-    def test_scan_help_with_group(self):
+    def test_GroupedHelpCommand_help_with_group(self):
         from plugincode import CORE_GROUP
 
-        @click.command(name='scan', cls=ScanCommand)
+        @click.command(name='scan', cls=GroupedHelpCommand)
         @click.option('--opt', is_flag=True, help='Help text for option',
                       help_group=CORE_GROUP, cls=PluggableCommandLineOption)
         def scan(opt):
